chore: added changes to clean codes, ensure we only create one notion record (#44)
